/olli books allbooks not working - fix allbooks so it actually gives the books to the command sender - moved books commands to O2Books (trying to make Ollivanders2.java more sane)
